move key_split from diagnostics to utils
